+++ src/components/TablePage/index.ts
@@ -17,4 +17,4 @@ export { default as TableActionsBar } from "./TableActionsBar";
 export * from "./TablePageLayout";
 export { default as TablePageLayout } from "./TablePageLayout";
 export * from "./types";
-export {EditTitleForm} from "@/components/TablePage/EditComponents/EditTitleForm";
+export {EditTableDetailsForm} from "@/components/TablePage/EditComponents/EditTableDetailsForm";
